pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dopeagents"
version = "0.1.0"
description = "Production-grade AI agents with typed interfaces and sophisticated multi-step workflows"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "DopeAgents Contributors" },
]
keywords = [
    "ai",
    "agents",
    "llm",
    "workflow",
    "orchestration",
    "typed-interfaces",
    "production",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.12",
    "pydantic-settings>=2.12",
    "instructor>=1.14",
    "litellm>=1.80",
    "langgraph>=1.1.3,<2.0",
    "python-dotenv>=1.0",
    "packaging>=21.0",
    "typing-extensions>=4.12",
    "click>=8.1",
    "jsonschema>=4.23",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.optional-dependencies]
mcp = ["fastmcp>=3.0,<4.0"]
cache = ["diskcache>=5.6"]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=0.41b0",
]
langchain = ["langchain>=0.1"]
crewai = ["crewai>=1.0"]
autogen = ["autogen-agentchat>=0.4"]
providers = ["openai>=1.0"]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]
lint = [
    "ruff>=0.3",
    "black>=24.1",
    "isort>=5.13",
    "mypy>=1.8",
]
all = [
    "fastmcp>=3.0,<4.0",
    "diskcache>=5.6",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=0.41b0",
    "langchain>=0.1",
    "crewai>=1.0",
    "autogen-agentchat>=0.4",
    "openai>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.3",
    "black>=24.1",
    "isort>=5.13",
    "pre-commit>=3.0",
]

[project.scripts]
dopeagents = "dopeagents.cli:main"

[project.urls]
Homepage = "https://github.com/cosmos-127/dopeagents"
Documentation = "https://docs.dopeagents.dev"
Repository = "https://github.com/cosmos-127/dopeagents"
Issues = "https://github.com/cosmos-127/dopeagents/issues"

[tool.setuptools]
packages = ["dopeagents"]

[tool.setuptools.package-data]
dopeagents = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [
    "docs",
    "build",
    "dist",
    ".venv",
    ".git",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by ruff format)
    "ARG001", # unused method argument (false positives)
    "B008",   # function call in default argument
    "I001",   # import sorting conflicts with isort - isort is the source of truth
]

[tool.ruff.lint.isort]
known-first-party = ["dopeagents"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-between-types = 1

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG001", "ARG002"]

[tool.mypy]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
explicit_package_bases = true
exclude = [
    ".venv",
    "build",
    "dist",
    "docs",
    "dopeagents.egg-info",
]

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "instructor.*",
    "litellm.*",
    "fastmcp.*",
    "opentelemetry.*",
    "pyautogen.*",
    "crewai.*",
    "openai.*",
    "diskcache.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["dopeagents"]
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
